slight cleanup of imports, more to do
