+++ packages/subgraph-gnosis-chain/tests/factory.test.ts
@@ -8,20 +8,22 @@ import {
 import { Address, BigInt, Bytes, ethereum } from "@graphprotocol/graph-ts";
 import { AMMPair, User } from "../generated/schema";
 import { log, newMockEvent } from "matchstick-as";
-import { ADDRESS_ZERO, GNO_ADDRESS, gno } from "../src/helpers";
-import { PairCreated } from "../generated/Factory/Factory";
-import { handleNewPair } from "../src/factory";
 import {
+  ADDRESS_ZERO,
+  GNO_ADDRESS,
+  gno,
   user1,
   user2,
   value,
   mockPair,
   value2x,
   data,
-  createPairCreatedEvent,
-} from "./helpers";
-import { Pair } from "../generated/templates/Pair/Pair";
+} from "../src/helpers";
+import { PairCreated } from "../generated/Factory/Factory";
+import { handleNewPair } from "../src/factory";
+import { createPairCreatedEvent } from "./helpers";
 
+// mock pair.totalSupply()
 createMockedFunction(
   Address.fromString(mockPair),
   "totalSupply",
@@ -30,11 +32,10 @@ createMockedFunction(
   .withArgs([])
   .returns([ethereum.Value.fromI32(100)]);
 
-Pair.bind(Address.fromString(mockPair));
-
-createMockedFunction(GNO_ADDRESS, "balanceOf", "balanceOf(address):(uint256)")
-  .withArgs([ethereum.Value.fromString(mockPair)])
-  .returns([ethereum.Value.fromI32(200)]);
+// // mock gno.balanceOf(pair.address)
+// createMockedFunction(GNO_ADDRESS, "balanceOf", "balanceOf(address):(uint256)")
+//   .withArgs([ethereum.Value.fromString(mockPair)])
+//   .returns([ethereum.Value.fromI32(200)]);
 
 test("Factory spawns pair", () => {
   clearStore();
